Fix #2: description always enabled
The browser doesn’t send description=off in the URL when the
description wasn’t checked, so we shouldn’t expect that. To make sure
that we still enable the description checkbox by default (that is, on
loading index.php without any query parameters), we extract the “do we
have form data” condition and use it for the description default too.