Don’t crash if no revisions are found
This fixes the underlying issue of #1, which was still reproducible
by using language codes that don’t exist: if the first SQL query
returns no results, we would try to run an ORES query like

    select … from ores_classification where … AND oresc_rev in ()

which is a syntax error. Showing an empty table in that case isn’t
ideal either (some feedback, e. g. “no revisions found, is there a
mistake in the language code?”, would be nice), but at least it’s not
an error anymore.