Use more efficient _recentchanges views
The default actor and comment views have to check against 7 other tables
whether an actor or a comment should be visible or not, which makes them
fairly inefficient. Since we only join against the recentchanges table,
we can use the more efficient _recentchanges views, which only check
whether a row is visible through recentchanges or not. For more
information, see the announcement [1] and the documentation [2].

[1]: https://lists.wikimedia.org/pipermail/cloud-announce/2019-June/000180.html
[2]: https://wikitech.wikimedia.org/wiki/News/Actor_storage_changes_on_the_Wiki_Replicas#The_actor_table_seems_really_slow--so_does_comment